Put person_info mehtod in different class
