+++ peopleInfo.rb
@@ -1,0 +1,10 @@
+class PeopleInfo
+def self.person_info(app)
+    puts 'What is the person\'s id?'
+    app.people.each_with_index do |person, index|
+        puts "#{index}. Name: #{person.name} ID: #{person.id} Age: #{person.age}"
+    end
+    id = gets.chomp
+    app.all_rentals(id.to_i)
+    end
+end
